modify float type columns to decimal type

+++ db/migrate/20220630074917_add_total_shares_to_asset.rb
@@ -1,5 +1,5 @@
 class AddTotalSharesToAsset < ActiveRecord::Migration[6.1]
   def change
-    add_column :assets, :total_shares, :float, default: 0
+    add_column :assets, :total_shares, :decimal, default: 0
   end
 end
